Apply code quality improvements from review comments
- Pass string and vector parameters by const reference to improve performance
- Fix signed/unsigned comparison warnings using size_t for loop variables
- Add error handling with try-catch for parseInt function
- Remove unnecessary object copying in TicketDAL::saveTickets
- Improve variable naming (tickets -> ticketItem, updateSeat -> seat)
- Fix .vscode/launch.json to use workspace-relative paths

+++ bus/BUSUtils.cpp
@@ -5,7 +5,7 @@
 
 std::string CurrencyFormat::formatCurrency(long long amount) {
     std::string money = std::to_string(amount);
-    for(int i=money.length()-3; i>0; i-=3) {
+    for(int i=static_cast<int>(money.length())-3; i>0; i-=3) {
         money.insert(i, ".");
     }
